feat(workshop): ui workshop library is a go

+++ libs/workshop/src/build.tsx
@@ -11,7 +11,6 @@ import { BuildArgs, HandlerCallback } from './types.js'
 import { Page } from './page.js'
 import { removeLeadingSlash, LIVE_RELOAD, createTmpDir } from './utils.js'
 
-
 export const build = ({ 
   exts,
   reload,
@@ -25,33 +24,39 @@ export const build = ({
   handlers: Map<string, HandlerCallback>
 ) => {
   const entryPoints = await fg(path.join(srcDir, `**/*${exts.startsWith('.') ? exts : '.' + exts}`))
-  const bundles = await bundler({
-    srcDir,
-    entryPoints,
-    reload,
-  })
-
-  // Create routes for bundled js
-  for(const bundle of bundles) {
-    const route = bundle[0]
-    handlers.set(route, (_: Request, res: Response) => {
-      res.setHeader('Content-Type', 'application/javascript')
-      res.setHeader('Content-Disposition', `attachment; filename=${path.basename(route)}`)
-      res.send(Buffer.from(bundle[1]))
+  let bundles: Map<string, Uint8Array>
+  try {
+    bundles = await bundler({
+      srcDir,
+      entryPoints,
+      reload,
     })
+  } catch (error){
+    console.error(error)
   }
-  
-  // create temp directory to write bundles to get story map
-  const [ formattedEntries, tempDirectory ] = await createTmpDir({ entryPoints, bundles, srcDir })
-  
-  // Get story map
-  const storyMap = await getStoryMap(formattedEntries, tempDirectory)
-  
-  // Clean up tmp directory
-  await fs.rm(tempDirectory, { recursive:true })
+  if(bundles.size) {
 
-  // Apply default styles
-  const [ _, stylesheet ] = css`
+    // Create routes for bundled js
+    for(const bundle of bundles) {
+      const route = bundle[0]
+      handlers.set(route, (_: Request, res: Response) => {
+        res.setHeader('Content-Type', 'application/javascript')
+        res.setHeader('Content-Disposition', `attachment; filename=${path.basename(route)}`)
+        res.send(Buffer.from(bundle[1]))
+      })
+    }
+  
+    // create temp directory to write bundles to get story map
+    const [ formattedEntries, tempDirectory ] = await createTmpDir({ entryPoints, bundles, srcDir })
+  
+    // Get story map
+    const storyMap = await getStoryMap(formattedEntries, tempDirectory)
+  
+    // Clean up tmp directory
+    await fs.rm(tempDirectory, { recursive:true })
+
+    // Apply default styles
+    const [ _, stylesheet ] = css`
   html {
     font-size: ${baseFontSize}px;
   }
@@ -66,25 +71,26 @@ export const build = ({
   ${transformCssTokens({ tokens, baseFontSize })}
   `
 
-  // Add story routes
-  for(const [ id, { clientPath, template: Template, attrs } ] of storyMap) {
-    handlers.set(`/${id}`, (_: Request, res: Response) => {
-      return res.send(ssr(<Page 
-        id={id}
-        reload={reload}
-        {...stylesheet}
-        clientPath={clientPath}
-      ><Template {...attrs} /></Page>))
-    })
-  }
+    // Add story routes
+    for(const [ id, { clientPath, template: Template, attrs } ] of storyMap) {
+      handlers.set(`/${id}`, (_: Request, res: Response) => {
+        return res.send(ssr(<Page 
+          id={id}
+          reload={reload}
+          {...stylesheet}
+          clientPath={clientPath}
+        ><Template {...attrs} /></Page>))
+      })
+    }
 
-  // Cleanup dead routes
-  for(const key of handlers.keys()) {
-    if(key === LIVE_RELOAD) continue
-    if(bundles.has(key) || storyMap.has(removeLeadingSlash(key))) continue
-    handlers.delete(key)
-  }
+    // Cleanup dead routes
+    for(const key of handlers.keys()) {
+      if(key === LIVE_RELOAD) continue
+      if(bundles.has(key) || storyMap.has(removeLeadingSlash(key))) continue
+      handlers.delete(key)
+    }
 
-  // Build playwright test
-  await writePlaywrightTests({ storyMap, testDir, srcDir, port, protocol })
+    // Build playwright test
+    await writePlaywrightTests({ storyMap, testDir, srcDir, port, protocol })
+  }
 }
